+++ CMAKE/lib/src/divider.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include "../inc/divider.h"
 
-int alex_divide(int a, int b){
-    int quotient = a / b;
+double alex_divide(int a, int b){
+    double quotient = (double)a / b;
     return quotient;
 }
